Move video capturing classes to 'media' package
