+++ webrtc-kmp/src/androidMain/kotlin/com/shepeliev/webrtckmp/MediaDevices.kt
@@ -5,6 +5,8 @@ package com.shepeliev.webrtckmp
 import android.Manifest
 import android.content.pm.PackageManager
 import androidx.core.content.ContextCompat
+import com.shepeliev.webrtckmp.media.CameraPermissionException
+import com.shepeliev.webrtckmp.media.CameraVideoCaptureController
 import org.webrtc.Camera2Enumerator
 import org.webrtc.MediaConstraints
 import java.util.UUID
